+++ lib/core/theme/app_radii.dart
@@ -47,3 +47,12 @@ class AppRadii extends ThemeExtension<AppRadii> {
     );
   }
 }
+
+const defaultAppRadii = AppRadii(
+  xs: 4,
+  s: 8,
+  m: 16,
+  l: 24,
+  xl: 32,
+  pill: 9999,
+);
