+++ lib/core/theme/app_spacing.dart
@@ -42,3 +42,11 @@ class AppSpacing extends ThemeExtension<AppSpacing> {
     );
   }
 }
+
+const defaultAppSpacing = AppSpacing(
+  xs: 4,
+  s: 8,
+  m: 16,
+  l: 24,
+  xl: 32,
+);
